+++ packages/react-wrap-with/src/index.ts
@@ -1,13 +1,15 @@
-import Extract from './Extract';
-import Spy from './Spy';
-import withProps from './withProps';
-import wrapWith from './wrapWith';
+// Related to https://github.com/import-js/eslint-plugin-import/issues/2872.
+// eslint-disable-next-line import/consistent-type-specifier-style
+import type { ConditionalKeys, Simplify } from 'type-fest';
+
+import Extract from './Extract.ts';
+import Spy from './Spy.ts';
+import withProps from './withProps.ts';
+import wrapWith from './wrapWith.ts';
 
 // Due to [a TypeScript bug with monorepo](https://github.com/microsoft/TypeScript/issues/47663), we need to export types from dependencies.
-import type { ConditionalKeys, Simplify } from 'type-fest';
-import type { HowOf } from './HowOf';
-import type { PropsOf } from './PropsOf';
-import type { RefOf } from './RefOf';
+import { type HowOf } from './HowOf.ts';
+import { type PropsOf } from './PropsOf.ts';
+import { type RefOf } from './RefOf.ts';
 
-export { ConditionalKeys, Extract, Simplify, Spy, withProps, wrapWith };
-export { type HowOf, type PropsOf, type RefOf };
+export { ConditionalKeys, Extract, Simplify, Spy, withProps, wrapWith, type HowOf, type PropsOf, type RefOf };